Fixes related to compiler warnings
Fix integer mismatch usages/comparisons.

Use "size_t" instead of "unsigned int" for data structures.

+++ src/archiver.h
@@ -32,9 +32,9 @@ typedef struct SDArchiverState {
   const SDArchiverParsed *parsed;
   FILE *out_f;
   SDArchiverHashMap *map;
-  unsigned int count;
-  unsigned int max;
-  unsigned int digits;
+  size_t count;
+  size_t max;
+  size_t digits;
 } SDArchiverState;
 
 enum SDArchiverStateReturns {
