+++ src/helpers.h
@@ -50,10 +50,10 @@ int simple_archiver_helper_make_dirs(const char *file_path);
 /// Returns non-NULL on success.
 /// Must be free'd with "free()" if non-NULL.
 /// start_idx is inclusive and end_idx is exclusive.
-char *simple_archiver_helper_cut_substr(const char *s, unsigned int start_idx,
-                                        unsigned int end_idx);
+char *simple_archiver_helper_cut_substr(const char *s, size_t start_idx,
+                                        size_t end_idx);
 
-unsigned int simple_archiver_helper_num_digits(unsigned int value);
+size_t simple_archiver_helper_num_digits(size_t value);
 
 void simple_archiver_helper_cleanup_FILE(FILE **fd);
 void simple_archiver_helper_cleanup_malloced(void **data);
