+++ src/parser_internal.h
@@ -21,7 +21,7 @@
 
 #include "parser.h"
 
-unsigned int simple_archiver_parser_internal_get_first_non_current_idx(
+size_t simple_archiver_parser_internal_get_first_non_current_idx(
     const char *filename);
 
 #endif
